requirements.txt
pydantic>=2.0
python-dotenv>=1.0
groq>=0.11.0
streamlit>=1.32
